Working on resending confirmation email, can no longer delete yourself (as an employee)

+++ routes/webapp/business/addemployees.js
@@ -122,33 +122,44 @@ exports.post = function(req,res){
             /*password: pass*/ //will be added programmatically once the employee confirms
         });
 
-        var app = require('../../../app');
-        var registrationLink;
-        if (app.get('env') == 'production') {
-            registrationLink = 'http://quart30.herokuapp.com/employeeregister?token=' + token;
-        }
-        else {
-            registrationLink = 'http://localhost:4000/employeeregister?token=' + token
-        }
-
-        var message = {
-            to: email,
-            from: 'quart30dev@gmail.com',
-            subject: 'Employee Signup',
-            text: 'Hello, ' + fname + ' ' + lname + ',\n\n' + 'Please click on the following link, or paste this into your browser to complete sign-up the process: \n\n' +
-                registrationLink
-        };
-
-        // send mail with defined transport object
-        transporter.sendMail(message, function(error, info){
-            if(error){
-                return console.log('Email error: ' + error);
-            }
-            console.log('Confirmation email sent: ' + info.response);
-        });
+        sendMail(fname, lname, email);
     }
     res.redirect('/addemployees');
 };
+
+
+/**
+ * Sends an email using nodemailer (registration link)
+ * @param fname first name
+ * @param lname last name
+ * @param email email
+ */
+function sendEmail(fname, lname, email) {
+    var app = require('../../../app');
+    var registrationLink;
+    if (app.get('env') == 'production') {
+        registrationLink = 'http://quart30.herokuapp.com/employeeregister?token=' + token;
+    }
+    else {
+        registrationLink = 'http://localhost:4000/employeeregister?token=' + token
+    }
+
+    var message = {
+        to: email,
+        from: 'quart30dev@gmail.com',
+        subject: 'Employee Signup',
+        text: 'Hello, ' + fname + ' ' + lname + ',\n\n' + 'Please click on the following link, or paste this into your browser to complete sign-up the process: \n\n' +
+        registrationLink
+    };
+
+    // send mail with defined transport object
+    transporter.sendMail(message, function(error, info){
+        if(error){
+            return console.log('Email error: ' + error);
+        }
+        console.log('Confirmation email sent: ' + info.response);
+    });
+}
 
 /** Deletes an employee from the database
  *
@@ -161,11 +172,14 @@ exports.delete = function (req, res) {
     var params = req.query;
     var email = params.email;
 
-    employeeDB.findOne({business: bid, email: email}, function (err, result) {
-        if (result) {
-            employeeDB.remove(result, {justOne: true});
-        }
-    });
+    // so you don't delete yourself
+    if (email !== req.user[0].email) {
+        employeeDB.findOne({business: bid, email: email}, function (err, result) {
+            if (result) {
+                employeeDB.remove(result, {justOne: true});
+            }
+        });
+    }
 
     res.redirect('/addemployees');
 };
